Move spherical2tilt into StylusData
